+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix invalid values in channel picker - #2313 by @orzechdev
 - Fix missing metadata and payment balance on unconfirmed orders - #2314 by @orzechdev
 - Fix exit form dialog false positive - #2311 by @orzechdev
+- Handle form errors before product creation - #2299 by @orzechdev
 - Fix no product error on unconfirmed order lines - #2324 by @orzechdev
 
 ## 3.4